# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Routes, note sets and notes for the RecceLog application.",
        "title": "RecceLog API",
        "contact": {
            "name": "RecceLog"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/notes/{setID}": {
            "get": {
                "description": "Returns the notes belonging to the given set, ordered by their `order` field.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get notes of a note set",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note set ID (UUID)",
                        "name": "setID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid note set id",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Note set not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the note set; ON DELETE CASCADE removes its notes.",
                "tags": [
                    "notes"
                ],
                "summary": "Delete a note set",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note set ID (UUID)",
                        "name": "setID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid note set id",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Not the note set author",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Note set not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
        "/v1/notes/{setID}/{noteID}": {
            "delete": {
                "description": "Deletes a note belonging to the given set. Returns 404 if the note does not exist or is not in this set.",
                "tags": [
                    "notes"
                ],
                "summary": "Delete a single note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note set ID (UUID)",
                        "name": "setID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note ID (UUID)",
                        "name": "noteID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid IDs",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Not the note set author",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Note not found (or not in this set)",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            },
            "patch": {
                "description": "Replaces position, order, type, severity, direction and description of a note in a given set.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Update a single note",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Note set ID (UUID)",
                        "name": "setID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Note ID (UUID)",
                        "name": "noteID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated note payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.NoteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid IDs or malformed JSON",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Not the note set author",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "404": {
                        "description": "Note not found (or not in this set)",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Problem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_RecceLog_api_internal_http_dto.Coordinate": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
...
        "github_com_RecceLog_api_internal_http_dto.NoteIn": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "position": {
                    "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.Coordinate"
                },
                "severity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_RecceLog_api_internal_http_dto.NoteResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "position": {
                    "$ref": "#/definitions/github_com_RecceLog_api_internal_http_dto.Coordinate"
                },
                "severity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_RecceLog_api_internal_http_problem.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "violations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_RecceLog_api_internal_http_problem.Violation"
                    }
                }
            }
        },
        "github_com_RecceLog_api_internal_http_problem.Violation": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}